requirements.txt
seaborn==0.11.1
streamlit==0.86.0
matplotlib==3.3.4
pandas==1.2.4
numpy==1.21.2
Pillow==8.3.2
